+++ src/api/v4/dataview/time-series/index.js
@@ -182,11 +182,15 @@ TimeSeries.prototype._createInternalModel = function (engine) {
     offset: this._offset,
     localTimezone: this._localTimezone,
     sync_on_bbox_change: !!this._boundingBoxFilter,
+    sync_on_circle_change: !!this._circleFilter,
+    sync_on_polygon_change: !!this._polygonFilter,
     enabled: this._enabled,
     column_type: 'date'
   }, {
     engine: engine,
-    bboxFilter: this._boundingBoxFilter && this._boundingBoxFilter.$getInternalModel()
+    bboxFilter: this._boundingBoxFilter && this._boundingBoxFilter.$getInternalModel(),
+    circleFilter: this._circleFilter && this._circleFilter.$getInternalModel(),
+    polygonFilter: this._polygonFilter && this._polygonFilter.$getInternalModel()
   });
 };