+++ src/api/v4/filter/base.js
@@ -27,3 +27,19 @@ module.exports = Base;
  * @type {carto.filter.Bounds}
  * @api
  */
+
+/**
+ * Fired when circle filter has changed. Handler gets a parameter with the new circle.
+ *
+ * @event circleChanged
+ * @type {carto.filter.CircleData}
+ * @api
+ */
+
+/**
+ * Fired when polygon filter has changed. Handler gets a parameter with the new polygon.
+ *
+ * @event polygonChanged
+ * @type {carto.filter.PolygonData}
+ * @api
+ */