+++ src/api/v4/filter/index.js
@@ -1,6 +1,8 @@
 const BoundingBox = require('./bounding-box');
 const BoundingBoxLeaflet = require('./bounding-box-leaflet');
 const BoundingBoxGoogleMaps = require('./bounding-box-gmaps');
+const Circle = require('./circle');
+const Polygon = require('./polygon');
 const Category = require('./category');
 const Range = require('./range');
 const AND = require('./and');
@@ -14,6 +16,8 @@ module.exports = {
   BoundingBox,
   BoundingBoxLeaflet,
   BoundingBoxGoogleMaps,
+  Circle,
+  Polygon,
   Category,
   Range,
   AND,