+++ src/api/v4/filter/spatial-filter-types.js
@@ -1,0 +1,10 @@
+/**
+ * Types of spatial filters
+ */
+var types = {
+  BBOX: 'bbox',
+  CIRCLE: 'circle',
+  POLYGON: 'polygon'
+};
+
+module.exports = types;